Move LibrariesService to services package
